requirements.txt
asgiref==3.5.2
Cerberus==1.3.2
certifi==2022.9.24
chardet==3.0.4
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
greenlet==2.0.0
gunicorn==20.0.4
idna==2.8
inflect==6.0.2
mysqlclient==2.0.3
numpy==1.21.3
pydantic==1.10.2
PyJWT==1.7.1
PyMySQL==0.9.3
pytube==12.1.0
pytz==2022.6
requests==2.22.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.3
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.25.11
moviepy==1.0.3